requirements.txt
# requirements.txt records the full set of dependencies for development
accelerate
codetiming
datasets
dill
hydra-core
liger-kernel
numpy
pandas
peft
pyarrow>=19.0.0
pybind11
pylatexenc
pre-commit
ray[default]
tensordict<=0.6.2
torchdata
transformers==4.51.3
packaging>=20.0
uvicorn
fastapi
scikit-learn==1.6.1
scipy==1.13.1
deepspeed
trl==0.18.1
volcengine-python-sdk[ark]
httpx==0.23.0
jsonlines
protobuf==3.20.*
vllm==0.8.3 
ray